+++ src/main/java/com/rmaslov/blog/photo/service/PhotoApiService.java
@@ -4,7 +4,12 @@ import com.mongodb.BasicDBObject;
 import com.mongodb.DBObject;
 import com.mongodb.client.gridfs.model.GridFSFile;
 import com.rmaslov.blog.album.exception.AlbumNotExistException;
+import com.rmaslov.blog.album.model.AlbumDoc;
 import com.rmaslov.blog.album.repository.AlbumRepository;
+import com.rmaslov.blog.auth.exceptions.AuthException;
+import com.rmaslov.blog.auth.exceptions.NotAccessException;
+import com.rmaslov.blog.auth.service.AuthService;
+import com.rmaslov.blog.base.service.CheckAccess;
 import com.rmaslov.blog.photo.api.request.PhotoSearchRequest;
 import com.rmaslov.blog.photo.mapping.PhotoMapping;
 import com.rmaslov.blog.base.api.request.SearchRequest;
@@ -15,6 +20,7 @@ import com.rmaslov.blog.photo.exception.PhotoNotExistException;
 import com.rmaslov.blog.photo.model.PhotoDoc;
 import com.rmaslov.blog.photo.repository.PhotoRepository;
 import com.rmaslov.blog.user.exception.UserNotExistException;
+import com.rmaslov.blog.user.model.UserDoc;
 import com.rmaslov.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
@@ -34,17 +40,20 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class PhotoApiService {
+public class PhotoApiService extends CheckAccess<PhotoDoc> {
     private final PhotoRepository photoRepository;
     private final MongoTemplate mongoTemplate;
     private final AlbumRepository albumRepository;
-    private final UserRepository userRepository;
     private final GridFsTemplate gridFsTemplate;
     private final GridFsOperations operations;
+    private final AuthService authService;
 
-    public PhotoDoc create(MultipartFile file, ObjectId ownerId, ObjectId albumId) throws PhotoExistException, UserNotExistException, AlbumNotExistException, IOException {
-        if(albumRepository.findById(albumId).isEmpty()) throw new AlbumNotExistException();
-        if(userRepository.findById(ownerId).isEmpty()) throw new UserNotExistException();
+    public PhotoDoc create(MultipartFile file,ObjectId albumId) throws AlbumNotExistException, IOException, AuthException, NotAccessException {
+        UserDoc userDoc = authService.currentUser();
+        AlbumDoc albumDoc = albumRepository.findById(albumId).orElseThrow(AlbumNotExistException::new);
+
+        if(albumDoc.getOwnerId().equals(userDoc.getId()) == false) throw new NotAccessException();
+
 
         DBObject metaData = new BasicDBObject();
         metaData.put("type", file.getContentType());
@@ -59,7 +68,7 @@ public class PhotoApiService {
                 .id(id)
                 .albumId(albumId)
                 .title(file.getOriginalFilename())
-                .ownerId(ownerId)
+                .ownerId(userDoc.getId())
                 .contentType(file.getContentType())
                 .build();
 
@@ -98,15 +107,16 @@ public class PhotoApiService {
         return SearchResponse.of(photoDocs, count);
     }
 
-    public PhotoDoc update(PhotoRequest request) throws PhotoNotExistException {
+    public PhotoDoc update(PhotoRequest request) throws PhotoNotExistException, AuthException, NotAccessException {
         Optional<PhotoDoc> photoDocOptional = photoRepository.findById(request.getId());
         if(photoDocOptional.isPresent() == false){
             throw new PhotoNotExistException();
         }
 
         PhotoDoc oldDoc = photoDocOptional.get();
+        UserDoc owner = checkAccess(oldDoc);
 
-        PhotoDoc photoDoc = PhotoMapping.getInstance().getRequest().convert(request);
+        PhotoDoc photoDoc = PhotoMapping.getInstance().getRequest().convert(request, owner.getId());
         photoDoc.setId(request.getId());
         photoDoc.setAlbumId(oldDoc.getAlbumId());
         photoDoc.setOwnerId(oldDoc.getOwnerId());
@@ -116,8 +126,19 @@ public class PhotoApiService {
         return photoDoc;
     }
 
-    public void delete(ObjectId id){
+    public void delete(ObjectId id) throws AuthException, NotAccessException, ChangeSetPersister.NotFoundException {
+        checkAccess(photoRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         gridFsTemplate.delete(new Query(Criteria.where("_id").is(id)));
         photoRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(PhotoDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
