+++ src/main/java/com/rmaslov/blog/todoTask/service/TodoTaskApiService.java
@@ -1,5 +1,9 @@
 package com.rmaslov.blog.todoTask.service;
 
+import com.rmaslov.blog.auth.exceptions.AuthException;
+import com.rmaslov.blog.auth.exceptions.NotAccessException;
+import com.rmaslov.blog.auth.service.AuthService;
+import com.rmaslov.blog.base.service.CheckAccess;
 import com.rmaslov.blog.todoTask.api.request.TodoTaskSearchRequest;
 import com.rmaslov.blog.todoTask.mapping.TodoTaskMapping;
 import com.rmaslov.blog.base.api.request.SearchRequest;
@@ -10,9 +14,11 @@ import com.rmaslov.blog.todoTask.exception.TodoTaskNotExistException;
 import com.rmaslov.blog.todoTask.model.TodoTaskDoc;
 import com.rmaslov.blog.todoTask.repository.TodoTaskRepository;
 import com.rmaslov.blog.user.exception.UserNotExistException;
+import com.rmaslov.blog.user.model.UserDoc;
 import com.rmaslov.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
@@ -26,15 +32,16 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class TodoTaskApiService {
+public class TodoTaskApiService extends CheckAccess<TodoTaskDoc> {
     private final TodoTaskRepository todoTaskRepository;
     private final MongoTemplate mongoTemplate;
-    private final UserRepository userRepository;
+    private final AuthService authService;
 
-    public TodoTaskDoc create(TodoTaskRequest request) throws TodoTaskExistException, UserNotExistException {
-        if(userRepository.findById(request.getOwnerId()).isEmpty()) throw new UserNotExistException();
+    public TodoTaskDoc create(TodoTaskRequest request) throws  AuthException {
 
-        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequest().convert(request);
+        UserDoc userDoc = authService.currentUser();
+
+        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequest().convert(request, userDoc.getId());
         todoTaskRepository.save(todoTaskDoc);
         return todoTaskDoc;
     }
@@ -45,10 +52,10 @@ public class TodoTaskApiService {
 
     public SearchResponse<TodoTaskDoc> search(
              TodoTaskSearchRequest request
-    ){
-        if(request.getOwnerId() == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+    ) throws AuthException {
+        UserDoc userDoc = authService.currentUser();
 
-        Criteria criteria = Criteria.where("ownerId").is(request.getOwnerId());
+        Criteria criteria = Criteria.where("ownerId").is(userDoc.getId());
 
         if(request.getQuery() != null && request.getQuery()  != ""){
             criteria = criteria.orOperator(
@@ -66,15 +73,16 @@ public class TodoTaskApiService {
         return SearchResponse.of(todoTaskDocs, count);
     }
 
-    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNotExistException {
+    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNotExistException, AuthException, NotAccessException {
         Optional<TodoTaskDoc> todoTaskDocOptional = todoTaskRepository.findById(request.getId());
         if(todoTaskDocOptional.isPresent() == false){
             throw new TodoTaskNotExistException();
         }
 
         TodoTaskDoc oldDoc = todoTaskDocOptional.get();
+        UserDoc userDoc = checkAccess(oldDoc);
 
-        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequest().convert(request);
+        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequest().convert(request, userDoc.getId());
 
         todoTaskDoc.setId(request.getId());
         todoTaskDoc.setOwnerId(oldDoc.getOwnerId());
@@ -83,7 +91,18 @@ public class TodoTaskApiService {
         return todoTaskDoc;
     }
 
-    public void delete(ObjectId id){
+    public void delete(ObjectId id) throws AuthException, NotAccessException, ChangeSetPersister.NotFoundException {
+        checkAccess(todoTaskRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         todoTaskRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(TodoTaskDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
